+++ administrator/components/com_users/views/groups/view.html.php
@@ -2,16 +2,16 @@
 /**
  * @version		$Id$
  * @copyright	Copyright (C) 2005 - 2009 Open Source Matters, Inc. All rights reserved.
- * @copyright	Copyright (C) 2008 - 2009 JXtended, LLC. All rights reserved.
  * @license		GNU General Public License version 2 or later; see LICENSE.txt
  */
 
+// No direct access.
 defined('_JEXEC') or die;
 
 jimport('joomla.application.component.view');
 
 /**
- * The HTML Users groups view.
+ * View class for a list of user groups.
  *
  * @package		Joomla.Administrator
  * @subpackage	com_users
@@ -19,6 +19,10 @@ jimport('joomla.application.component.view');
  */
 class UsersViewGroups extends JView
 {
+	protected $state;
+	protected $items;
+	protected $pagination;
+
 	/**
 	 * Display the view
 	 */
@@ -38,28 +42,39 @@ class UsersViewGroups extends JView
 		$this->assignRef('items',		$items);
 		$this->assignRef('pagination',	$pagination);
 
-		parent::display($tpl);
 		$this->_setToolbar();
+		parent::display($tpl);
 	}
 
 	/**
-	 * Build the default toolbar.
-	 *
-	 * @return	void
+	 * Setup the Toolbar.
 	 */
 	protected function _setToolbar()
 	{
+		$canDo	= UsersHelper::getActions();
+
 		JToolBarHelper::title(JText::_('Users_View_Groups_Title'), 'groups');
 
-		JToolBarHelper::custom('group.add', 'new.png', 'new_f2.png', 'New', false);
-		JToolBarHelper::custom('group.edit', 'edit.png', 'edit_f2.png', 'Edit', true);
-		JToolBarHelper::deleteList('', 'group.delete');
+		if ($canDo->get('core.create'))
+		{
+			JToolBarHelper::custom('group.add', 'new.png', 'new_f2.png', 'New', false);
+		}
+		if ($canDo->get('core.edit'))
+		{
+			JToolBarHelper::custom('group.edit', 'edit.png', 'edit_f2.png', 'Edit', true);
+		}
+		if ($canDo->get('core.delete'))
+		{
+			JToolBarHelper::deleteList('', 'group.delete');
+		}
 
 		JToolBarHelper::divider();
 
-		// We can't use the toolbar helper here because there is no generic popup button.
-		$bar = &JToolBar::getInstance('toolbar');
-		$bar->appendButton('Popup', 'config', 'JToolbar_Options', 'index.php?option=com_users&view=config&tmpl=component', 570, 500);
+		if ($canDo->get('core.admin'))
+		{
+			JToolBarHelper::preferences('com_users');
+		}
+		JToolBarHelper::divider();
 		JToolBarHelper::help('screen.users.groups');
 	}
 }
